Module aws_sdk_servicediscovery::error

Modules

custom_health_not_found

See CustomHealthNotFound

duplicate_request

See DuplicateRequest

instance_not_found

See InstanceNotFound

invalid_input

See InvalidInput

namespace_already_exists

See NamespaceAlreadyExists

namespace_not_found

See NamespaceNotFound

operation_not_found

See OperationNotFound

request_limit_exceeded

See RequestLimitExceeded

resource_in_use

See ResourceInUse

resource_limit_exceeded

See ResourceLimitExceeded

resource_not_found_exception

See ResourceNotFoundException

service_already_exists

See ServiceAlreadyExists

service_not_found

See ServiceNotFound

too_many_tags_exception

See TooManyTagsException

Structs

CreateHttpNamespaceError

CreatePrivateDnsNamespaceError

CreatePublicDnsNamespaceError

CreateServiceError

CustomHealthNotFound

The health check for the instance that's specified by ServiceId and InstanceId isn't a

DeleteNamespaceError

DeleteServiceError

DeregisterInstanceError

DiscoverInstancesError

DuplicateRequest

The operation is already in progress.

GetInstanceError

GetInstancesHealthStatusError

GetNamespaceError

GetOperationError

GetServiceError

InstanceNotFound

No instance exists with the specified ID, or the instance was recently registered, and information about the

InvalidInput

One or more specified values aren't valid. For example, a required value might be missing, a numeric value might

ListInstancesError

ListNamespacesError

ListOperationsError

ListServicesError

ListTagsForResourceError

NamespaceAlreadyExists

The namespace that you're trying to create already exists.

NamespaceNotFound

No namespace exists with the specified ID.

OperationNotFound

No operation exists with the specified ID.

RegisterInstanceError

RequestLimitExceeded

The operation can't be completed because you've reached the quota for the number of requests. For more

ResourceInUse

The specified resource can't be deleted because it contains other resources. For example, you can't delete a

ResourceLimitExceeded

The resource can't be created because you've reached the quota on the number of resources.

ResourceNotFoundException

The operation can't be completed because the resource was not found.

ServiceAlreadyExists

The service can't be created because a service with the same name already exists.

ServiceNotFound

No service exists with the specified ID.

TagResourceError

TooManyTagsException

The list of tags on the resource is over the quota. The maximum number of tags that can be applied to a resource

UntagResourceError

UpdateHttpNamespaceError

UpdateInstanceCustomHealthStatusError

UpdatePrivateDnsNamespaceError

UpdatePublicDnsNamespaceError

UpdateServiceError

Enums

CreateHttpNamespaceErrorKind

CreatePrivateDnsNamespaceErrorKind

CreatePublicDnsNamespaceErrorKind

CreateServiceErrorKind

DeleteNamespaceErrorKind

DeleteServiceErrorKind

DeregisterInstanceErrorKind

DiscoverInstancesErrorKind

GetInstanceErrorKind

GetInstancesHealthStatusErrorKind

GetNamespaceErrorKind

GetOperationErrorKind

GetServiceErrorKind

ListInstancesErrorKind

ListNamespacesErrorKind

ListOperationsErrorKind

ListServicesErrorKind

ListTagsForResourceErrorKind

RegisterInstanceErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateHttpNamespaceErrorKind

UpdateInstanceCustomHealthStatusErrorKind

UpdatePrivateDnsNamespaceErrorKind

UpdatePublicDnsNamespaceErrorKind

UpdateServiceErrorKind