Module aws_sdk_servicediscovery::model

Modules

dns_config

See DnsConfig

dns_config_change

See DnsConfigChange

dns_properties

See DnsProperties

dns_record

See DnsRecord

health_check_config

See HealthCheckConfig

health_check_custom_config

See HealthCheckCustomConfig

http_instance_summary

See HttpInstanceSummary

http_namespace_change

See HttpNamespaceChange

http_properties

See HttpProperties

instance

See Instance

instance_summary

See InstanceSummary

namespace

See Namespace

namespace_filter

See NamespaceFilter

namespace_properties

See NamespaceProperties

namespace_summary

See NamespaceSummary

operation

See Operation

operation_filter

See OperationFilter

operation_summary

See OperationSummary

private_dns_namespace_change

See PrivateDnsNamespaceChange

private_dns_namespace_properties

See PrivateDnsNamespaceProperties

private_dns_namespace_properties_change

See PrivateDnsNamespacePropertiesChange

private_dns_properties_mutable

See PrivateDnsPropertiesMutable

private_dns_properties_mutable_change

See PrivateDnsPropertiesMutableChange

public_dns_namespace_change

See PublicDnsNamespaceChange

public_dns_namespace_properties

See PublicDnsNamespaceProperties

public_dns_namespace_properties_change

See PublicDnsNamespacePropertiesChange

public_dns_properties_mutable

See PublicDnsPropertiesMutable

public_dns_properties_mutable_change

See PublicDnsPropertiesMutableChange

service

See Service

service_change

See ServiceChange

service_filter

See ServiceFilter

service_summary

See ServiceSummary

soa

See Soa

soa_change

See SoaChange

tag

See Tag

Structs

DnsConfig

A complex type that contains information about the Amazon Route 53 DNS records that you want Cloud Map to create when

DnsConfigChange

A complex type that contains information about changes to the Route 53 DNS records that Cloud Map creates when you

DnsProperties

A complex type that contains the ID for the Route 53 hosted zone that Cloud Map creates when you create a

DnsRecord

A complex type that contains information about the Route 53 DNS records that you want Cloud Map to create when you

HealthCheckConfig

HealthCheckCustomConfig

A complex type that contains information about an optional custom health check. A custom health check, which

HttpInstanceSummary

In a response to a DiscoverInstances request, HttpInstanceSummary contains information about one instance that

HttpNamespaceChange

Updated properties for the HTTP

HttpProperties

A complex type that contains the name of an HTTP namespace.

Instance

A complex type that contains information about an instance that Cloud Map creates when you submit a

InstanceSummary

A complex type that contains information about the instances that you registered by using a specified

Namespace

A complex type that contains information about a specified namespace.

NamespaceFilter

A complex type that identifies the namespaces that you want to list. You can choose to list public or private

NamespaceProperties

A complex type that contains information that's specific to the namespace type.

NamespaceSummary

A complex type that contains information about a namespace.

Operation

A complex type that contains information about a specified operation.

OperationFilter

A complex type that lets you select the operations that you want to list.

OperationSummary

A complex type that contains information about an operation that matches the criteria that you specified in a

PrivateDnsNamespaceChange

Updated properties for the private DNS

PrivateDnsNamespaceProperties

DNS properties for the private DNS

PrivateDnsNamespacePropertiesChange

Updated properties for the private DNS

PrivateDnsPropertiesMutable

DNS properties for the private DNS

PrivateDnsPropertiesMutableChange

Updated DNS properties for the private

PublicDnsNamespaceChange

Updated properties for the public DNS

PublicDnsNamespaceProperties

DNS properties for the public DNS

PublicDnsNamespacePropertiesChange

Updated properties for the public DNS

PublicDnsPropertiesMutable

DNS properties for the public DNS

PublicDnsPropertiesMutableChange

Updated DNS properties for the public DNS

Service

A complex type that contains information about the specified service.

ServiceChange

A complex type that contains changes to an existing service.

ServiceFilter

A complex type that lets you specify the namespaces that you want to list services for.

ServiceSummary

A complex type that contains information about a specified service.

Soa

Start of Authority (SOA) properties for a

SoaChange

Updated Start of Authority (SOA)

Tag

A custom key-value pair that's associated with a resource.

Enums

CustomHealthStatus

FilterCondition

HealthCheckType

HealthStatus

NOTE: HealthStatus::Unknown has been renamed to ::UnknownValue.

HealthStatusFilter

NamespaceFilterName

NamespaceType

OperationFilterName

OperationStatus

OperationTargetType

OperationType

RecordType

RoutingPolicy

ServiceFilterName

ServiceType

ServiceTypeOption